# api/openapi.yaml
openapi: 3.0.1
info:
  version: 1.0.0
  title: Hacker News (ProjecteG14B)
  description: "This is the Open API documentation for the REST API of our beloved application **Hacker News** deployed at <https://hackernews-projecteg14b.herokuapp.com/>. <br>¡ATENCIÓN, ESTÁ PENDIENTE! All operations are executable. Only one operation requires authentication: `deleteTweet`. In this case, you must **Authorize** your request by providing the api_key vaule you got when you created the tweet."
servers:
  - url: 'https://hackernews-projecteg14b.herokuapp.com/'
  - url: 'http://54.87.189.232:8080'
paths:
  /contributions:
    get:
      tags:
      - contributions
      summary: Retrieves all the Contributions of type url (ordered by points), of type ask (ordered by points), of type all (ordered by creation), created by one User or voted by the logged in User
      operationId: getContributions
      parameters:
      - name: type
        in: query
        description: Contribution type to filter
        required: false
        schema:
          type: string
          enum: [all, url, ask]
      - name: user_id
        in: query
        description: Contribution user to filter
        schema:
          type: integer
      - name: upvotedS
        in: query
        description: upvoted contributions of logged in User
        schema:
          type: integer
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Contribution'
        404:
          description: 'Error: Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorUserExample:
                  $ref: '#/components/examples/error404User'
        401:
          description: 'Error: Unauthorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorMissingAuthenticationExample:
                  $ref: '#/components/examples/error401UnauthorizedMissingAuthentication'
                errorIncorrectAuthenticationExample:
                  $ref: '#/components/examples/error401UnauthorizedIncorrectAuthentication'
        403:
          description: 'Error: Forbidden'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorPrivilegeNotGrantedExample:
                  $ref: '#/components/examples/error403Forbidden'
    post:
      tags:
      - contributions
      summary: Adds a new Contribution to the HackerNews
      operationId: addContribution
      requestBody:
        description: Provide the title, the url (optional) & text (optional) of the new Contribution
        content:
          application/json:
            schema:
              required:
              - title
              - url
              - text
              type: object
              properties:
                title:
                  type: string
                  minLength: 1
                  example: 'Contribution title'
                url:
                  type: string
                  format: url
                  nullable: true
                  example: 'https://www.contribution-url.com'
                text:
                  type: string
                  nullable: true
                  example: 'Contribution text'
        required: true
      responses:
        201:
          description: Contribution created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contribution'
        400:
          description: 'Error: Bad Request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorTitleEmptyExample:
                  $ref: '#/components/examples/error400TitleContribution'
                errorURLInvalidExample:
                  $ref: '#/components/examples/error400URLContribution'
        401:
            description: 'Error: Unauthorized'
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResult'
                examples:
                  errorMissingAuthenticationExample:
                    $ref: '#/components/examples/error401UnauthorizedMissingAuthentication'
                  errorIncorrectAuthenticationExample:
                    $ref: '#/components/examples/error401UnauthorizedIncorrectAuthentication'
        409:
            description: 'Error: Conflict'
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResult'
                examples:
                  errorURLExistenceExample:
                    $ref: '#/components/examples/error409Contribution'
...
components:
  schemas:
    Contribution:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          minLength: 1
          example: 'Contribution title'
        url:
          type: string
          format: url
          nullable: true
          example: 'https://www.contribution-url.com'
        text:
          type: string
          nullable: true
          example: ''
        created_at:
          type: string
          format: date-time
        tipus:
          type: string
          example: 'url'
        points:
          type: integer
          example: 0
        user:
          type: string
          format: email
          example: 'user@email.com'
        user_id:
          type: integer
          example: 1
        #comments:
        #  type: array
        #  items:
        #    type: object
...
    ErrorResult:
      type: object
      properties:
        status:
          type: integer
          enum: [400, 401, 403, 404, 409]
        error: 
          type: string
          enum: ["Bad Request", "Unauthorized", "Forbidden", "Not Found", "Conflict"]
        message: 
          type: string
  examples:
    error404User:
      value:
        status: 404
        error: "Not Found"
        message: "No User with that user_id"
    error401UnauthorizedMissingAuthentication:
      value:
        status: 401
        error: "Unauthorized"
        message: "Missing authentication"
    error401UnauthorizedIncorrectAuthentication:
      value:
        status: 401
        error: "Unauthorized"
        message: "Incorrect authentication"
    error403Forbidden:
      value:
        status: 403
        error: "Forbidden"
        message: "Privilege not granted"
...
    error409Contribution:
      value:
        status: 409
        error: "Conflict"
        message: "URL already exists"
    error400TitleContribution:
      value:
        status: 400
        error: "Bad Request"
        message: "Title cannot be empty"
    error400URLContribution:
      value:
        status: 400
        error: "Bad Request"
        message: "URL must be valid"
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      name: X-API-KEY #Custom headers in rails apps must have the X- prefix
      in: header
security:
  - ApiKeyAuth: []